Remove the default date
There is a option of giving the date of today but it is saved once then never
changed: it is today only the first time it is saved in the template.  Moreover,
a user may do not want any due date in a task; the default value of date has
been removed and leaving this field empty is possible now.